Extracted bottom spaccing effect on hotelbar to a component
