+++ src/components/Hotel.css
@@ -12,12 +12,3 @@
 .hotel {
   position: relative;
 }
-
-.scrollbar-space {
-  position: relative;
-  width: 6rem;
-  background-color: white;
-  border: 2px solid black;
-  border-bottom: 0;
-  border-left: 0;
-}
