+++ src/components/Hotel.tsx
@@ -7,6 +7,7 @@ import * as RoomTypesSlice from "../redux/roomTypesSlice";
 
 import Floor from "./hotel/Floor";
 import RoomNumber from "./hotel/RoomNumber";
+import BottomSpace from "./hotel/BottomSpace";
 
 import "./Hotel.css";
 
@@ -14,14 +15,11 @@ function Hotel(): JSX.Element {
   const dispatch = useAppDispatch();
   const hotelData = useHotelData();
   const scrollTop = useScrollTop();
-  const offsetHeight = useTableOffsetHeight();
-  const clientHeight = useTableClientHeight();
   const rows = useRowsMemo(hotelData);
   const ref = useRef<HTMLDivElement>(null);
 
   useHotelDataFetchingEffect(dispatch);
   useRoomTypesFetchingEffect(dispatch);
-  useHotelbarBottomSpacingEffect(offsetHeight, clientHeight, rows);
   useScrollEffect(ref, scrollTop);
 
   return (
@@ -31,14 +29,6 @@ function Hotel(): JSX.Element {
       </div>
     </div>)
   ;
-}
-
-function useTableOffsetHeight() {
-  return useAppSelector(state => state.table.offsetHeight);
-}
-
-function useTableClientHeight() {
-  return useAppSelector(state => state.table.clientHeight);
 }
 
 function useScrollTop() {
@@ -59,6 +49,7 @@ function useRowsMemo(hotelData: HotelSlice.HotelData): JSX.Element[] {
         rows.push(<RoomNumber number={room.number} key={room.number} isLast={isLast} />);
       });
     });
+    rows.push(<BottomSpace key="bottom-space" />);
 
     return rows;
   }, [hotelData.floors]);
@@ -74,31 +65,6 @@ function useRoomTypesFetchingEffect(dispatch: React.Dispatch<RoomTypesSlice.Fetc
   useEffect(() => {
     dispatch(RoomTypesSlice.fetchAsync());
   }, [dispatch]);
-}
-
-function useHotelbarBottomSpacingEffect(
-  offsetHeight: number,
-  clientHeight: number,
-  rows: JSX.Element[]
-): void {
-  useEffect(() => {
-    const scrollbarWidth = offsetHeight - clientHeight - 1;
-    if (scrollbarWidth > 0) {
-      rows.push(
-        <div
-          className="scrollbar-space"
-          key="scrollbar-space"
-          style={{ height: scrollbarWidth }}>
-        </div>
-      );
-    }
-
-    return () => {
-      if (scrollbarWidth > 0) {
-        rows.pop();
-      }
-    };
-  }, [offsetHeight, clientHeight, rows]);
 }
 
 function useScrollEffect(ref: React.RefObject<HTMLDivElement>, scrollTop: number): void {
