+++ src/components/hotel/BottomSpace.css
@@ -1,0 +1,8 @@
+.scrollbar-space {
+  position: relative;
+  width: 6rem;
+  background-color: white;
+  border: 2px solid black;
+  border-bottom: 0;
+  border-left: 0;
+}
